% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.pycox-package}
\alias{mlr3learners.pycox}
\alias{mlr3learners.pycox-package}
\title{mlr3learners.pycox: Adds Learners from Python Package pycox for mlr3}
\description{
Adds Learners from Python Package pycox for mlr3.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
